requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
PyYAML==3.12
requests==2.18.4
urllib3==1.22
wheel
appdirs==1.4.3
